Refactor to index.js according to Gatsby structure, homepage loads
