+++ gatsby-config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  siteMetadata: {
+    title: `Nile Camai`,
+  },
+  plugins: [
+    `gatsby-plugin-sass`
+  ],
+}
