+++ package.json
@@ -8,24 +8,22 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "bootstrap": "^4.5.2",
+    "gatsby": "^4.8.1",
+    "gatsby-plugin-sass": "^5.8.0",
     "jquery": "^3.6.0",
     "popper.js": "^1.16.1",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-device-detect": "^1.14.0",
     "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.1",
     "react-scroll": "^1.8.2",
     "react-social-icons": "^4.1.0",
     "react-spring": "^8.0.27",
-    "sass": "^1.41.1"
+    "sass": "^1.49.9"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "gatsby develop -p 3333 -o",
+    "build": "gatsby build",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build"
   },
@@ -45,6 +43,7 @@
     ]
   },
   "devDependencies": {
+    "eslint-config-react-app": "^7.0.0",
     "gh-pages": "^3.1.0"
   }
 }
