+++ public/page-data/AboutPage/page-data.json
@@ -1,0 +1,5 @@
+{
+    "componentChunkName": "component---src-pages-about-page-js",
+    "path": "/AboutPage/",
+    "result": {"pageContext":{}},
+    "staticQueryHashes": []}
