+++ public/page-data/ContactPage/page-data.json
@@ -1,0 +1,5 @@
+{
+    "componentChunkName": "component---src-pages-contact-page-js",
+    "path": "/ContactPage/",
+    "result": {"pageContext":{}},
+    "staticQueryHashes": []}
