+++ public/page-data/HomePage/page-data.json
@@ -1,0 +1,5 @@
+{
+    "componentChunkName": "component---src-pages-home-page-js",
+    "path": "/HomePage/",
+    "result": {"pageContext":{}},
+    "staticQueryHashes": []}
