+++ public/page-data/ProjectsPage/page-data.json
@@ -1,0 +1,5 @@
+{
+    "componentChunkName": "component---src-pages-projects-page-js",
+    "path": "/ProjectsPage/",
+    "result": {"pageContext":{}},
+    "staticQueryHashes": []}
