+++ public/page-data/dev-404-page/page-data.json
@@ -1,0 +1,5 @@
+{
+    "componentChunkName": "component---cache-dev-404-page-js",
+    "path": "/dev-404-page/",
+    "result": {"data":{"allSiteFunction":{"nodes":[]},"allSitePage":{"nodes":[{"path":"/AboutPage/"},{"path":"/ContactPage/"},{"path":"/HomePage/"},{"path":"/ProjectsPage/"},{"path":"/"}]}},"pageContext":{}},
+    "staticQueryHashes": []}
